PyhtonApi/requirements.txt
Flask==2.3.3
Flask-CORS==3.0.10
pdfminer.six==20221105
spacy==3.6.0
pymongo==4.5.0
